+++ CHANGELOG.md
@@ -7,11 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Added `LenientMap.delete/2`.
+
 ## 1.2.0
 
 ### Added
 
-- Added `LenientMap.get/3`
+- Added `LenientMap.get/3`.
 
 ## 1.1.0